refactor: Aplicado o pattern para melhor entendimento dos códigos
