+++ lib/components/index.ts
@@ -1,0 +1,2 @@
+export * from '../components/_searchselect'
+export * from '../components/_select'
