+++ lib/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './_search'
+export * from './_data'
